requirements (1).txt
# Core ML
torch==2.3.0
torchvision==0.18.0

# Training utils
tqdm==4.66.4

# HuggingFace ecosystem
transformers==4.41.2
safetensors==0.4.3

# OpenCLIP (для CLIP энкодера)
open_clip_torch==2.24.0

# UI
gradio==4.36.1

# Utils
numpy==1.26.4
pillow==10.3.0
matplotlib==3.8.4
